#0: floor test

+++ tests/tt_eager/python_api_testing/unit_testing/backward_ops/utility_funcs.py
@@ -23,16 +23,16 @@ def data_gen_pt_tt(input_shapes, device, required_grad=False):
 def data_gen_with_range(input_shapes, low, high, device, required_grad=False, is_row_major=False):
     assert high > low, "Incorrect range provided"
     torch.manual_seed(213919)
-    pt_tensor = torch.rand(input_shapes, requires_grad=required_grad).bfloat16() * (high - low) + low
+    pt_tensor = torch.rand(input_shapes, requires_grad=required_grad).float() * (high - low) + low
     if is_row_major:
         tt_tensor = (
-            tt_lib.tensor.Tensor(pt_tensor, tt_lib.tensor.DataType.BFLOAT16)
+            tt_lib.tensor.Tensor(pt_tensor, tt_lib.tensor.DataType.FLOAT32)
             .to(tt_lib.tensor.Layout.ROW_MAJOR)
             .to(device)
         )
     else:
         tt_tensor = (
-            tt_lib.tensor.Tensor(pt_tensor, tt_lib.tensor.DataType.BFLOAT16).to(tt_lib.tensor.Layout.TILE).to(device)
+            tt_lib.tensor.Tensor(pt_tensor, tt_lib.tensor.DataType.FLOAT32).to(tt_lib.tensor.Layout.TILE).to(device)
         )
     return pt_tensor, tt_tensor
